# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenASM API",
        "description": "OpenASM (Open Asset Management) REST API for comprehensive asset management with multi-tenancy support",
        "contact": {
            "name": "OpenASM Support",
            "email": "admin@openasm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "OpenASM API Server"
        }
    ],
    "paths": {
...
        "/api/v1/organization": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get current organization details",
                "operationId": "4699fd1d00f8d6a5e5fbd390647f87ef",
                "responses": {
                    "200": {
                        "description": "Organization details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No organization assigned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organization/overview": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get current organization overview with detailed information",
                "operationId": "fec89867b66837747d0fc7e4d59766cf",
                "responses": {
                    "200": {
                        "description": "Organization overview retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrganizationResource"
                                        },
                                        "meta": {
                                            "properties": {
                                                "statistics": {
                                                    "type": "object"
                                                },
                                                "recent_customers": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CustomerResource"
                                                    }
                                                },
                                                "recent_assets": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AssetResource"
                                                    }
                                                },
                                                "health": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organization/statistics": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get current organization statistics",
                "operationId": "d1fb067b1c4a2f607fef8c19d7e04d64",
                "responses": {
                    "200": {
                        "description": "Organization statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "users_count": {
                                                    "type": "integer"
                                                },
                                                "customers_count": {
                                                    "type": "integer"
                                                },
                                                "assets_count": {
                                                    "type": "integer"
                                                },
                                                "locations_count": {
                                                    "type": "integer"
                                                },
                                                "active_assets_count": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string"
                    },
                    "error": {
                        "description": "Detailed error information",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CustomerResource": {
                "properties": {
                    "id": {
                        "description": "Customer ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "organization_id": {
                        "description": "Organization ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Customer name",
                        "type": "string"
                    },
                    "customer_code": {
                        "description": "Customer code",
                        "type": "string"
                    },
                    "email": {
                        "description": "Customer email",
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "description": "Customer phone",
                        "type": "string",
                        "nullable": true
                    },
                    "website": {
                        "description": "Customer website",
                        "type": "string",
                        "nullable": true
                    },
                    "address_line_1": {
                        "description": "Address line 1",
                        "type": "string",
                        "nullable": true
                    },
                    "address_line_2": {
                        "description": "Address line 2",
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "description": "City",
                        "type": "string",
                        "nullable": true
                    },
                    "state": {
                        "description": "State/Province",
                        "type": "string",
                        "nullable": true
                    },
                    "postal_code": {
                        "description": "Postal code",
                        "type": "string",
                        "nullable": true
                    },
                    "country": {
                        "description": "Country",
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Active status",
                        "type": "boolean"
                    },
                    "data_quality_score": {
                        "description": "Data quality score (0-100)",
                        "type": "integer"
                    },
                    "assets_count": {
                        "description": "Number of assets",
                        "type": "integer"
                    },
                    "metadata": {
                        "description": "Additional metadata",
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "OrganizationResource": {
                "title": "Organization Resource",
                "description": "Organization data representation",
                "properties": {
                    "id": {
                        "description": "Organization unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Organization name",
                        "type": "string"
                    },
                    "code": {
                        "description": "Organization code",
                        "type": "string"
                    },
                    "email": {
                        "description": "Organization email",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Organization phone",
                        "type": "string"
                    },
                    "website": {
                        "description": "Organization website",
                        "type": "string",
                        "format": "uri"
                    },
                    "description": {
                        "description": "Organization description",
                        "type": "string"
                    },
                    "address": {
                        "description": "Organization address",
                        "type": "string"
                    },
                    "city": {
                        "description": "Organization city",
                        "type": "string"
                    },
                    "state": {
                        "description": "Organization state",
                        "type": "string"
                    },
                    "country": {
                        "description": "Organization country",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Organization postal code",
                        "type": "string"
                    },
                    "is_active": {
                        "description": "Whether the organization is active",
                        "type": "boolean"
                    },
                    "metadata": {
                        "description": "Additional organization metadata",
                        "type": "object"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "AssetResource": {
                "properties": {
                    "id": {
                        "description": "Asset ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "organization_id": {
                        "description": "Organization ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "customer_id": {
                        "description": "Customer ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "location_id": {
                        "description": "Location ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "oem_id": {
                        "description": "OEM ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "product_id": {
                        "description": "Product ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "type_id": {
                        "description": "Type ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "status_id": {
                        "description": "Status ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "serial_number": {
                        "description": "Serial number",
                        "type": "string"
                    },
                    "asset_tag": {
                        "description": "Asset tag",
                        "type": "string",
                        "nullable": true
                    },
                    "model_number": {
                        "description": "Model number",
                        "type": "string",
                        "nullable": true
                    },
                    "part_number": {
                        "description": "Part number",
                        "type": "string",
                        "nullable": true
                    },
                    "name": {
                        "description": "Asset name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description",
                        "type": "string",
                        "nullable": true
                    },
                    "purchase_date": {
                        "description": "Purchase date",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "installation_date": {
                        "description": "Installation date",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "warranty_start_date": {
                        "description": "Warranty start date",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "warranty_end_date": {
                        "description": "Warranty end date",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "purchase_price": {
                        "description": "Purchase price",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "current_value": {
                        "description": "Current value",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Active status",
                        "type": "boolean"
                    },
                    "data_quality_score": {
                        "description": "Data quality score (0-100)",
                        "type": "integer"
                    },
                    "metadata": {
                        "description": "Additional metadata",
                        "type": "object",
                        "nullable": true
                    },
                    "warranty_status": {
                        "description": "Warranty status (active, expiring_soon, expired, no_warranty)",
                        "type": "string"
                    },
                    "has_active_warranty": {
                        "description": "Whether asset has active warranty",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "organization": {
                        "description": "Organization details",
                        "type": "object",
                        "nullable": true
                    },
                    "customer": {
                        "description": "Customer details",
                        "type": "object",
                        "nullable": true
                    },
                    "location": {
                        "description": "Location details",
                        "type": "object",
                        "nullable": true
                    },
                    "oem": {
                        "description": "OEM details",
                        "type": "object",
                        "nullable": true
                    },
                    "product": {
                        "description": "Product details",
                        "type": "object",
                        "nullable": true
                    },
                    "type": {
                        "description": "Asset type details",
                        "type": "object",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status details",
                        "type": "object",
                        "nullable": true
                    },
                    "warranties": {
                        "description": "Warranty records",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "tags": {
                        "description": "Associated tags",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}